only the string to fix

+++ utils/ft_putnbro_rtn_neg.c
@@ -6,7 +6,7 @@
 /*   By: pirichar <pirichar@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/10 10:08:03 by pirichar          #+#    #+#             */
-/*   Updated: 2021/11/10 12:52:50 by pirichar         ###   ########.fr       */
+/*   Updated: 2021/11/10 21:18:00 by pirichar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putcharo_rtn(char c, int *sum);
 
-void	ft_putnbro_rtn_neg(int nb, int *sum)
+void	ft_putnbro_rtn_neg(long nb, int *sum)
 {
 	long	n;
